requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
Django==3.1
django-graphql-jwt==0.3.1
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
